pyproject.toml
[project]
name = "ocr-webapp"
version = "0.1.0"
description = "A modular, production-ready OCR web application"
authors = [
    {name = "cytech"}
]
requires-python = "^3.11"
readme = "README.md"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.9",
    "pillow>=10.0.0",
    "pytesseract>=0.3.10",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
    "black>=24.0.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "UP", "B", "C4", "PTH"]
ignore = ["E501"]

[tool.ruff.per-file-ignores]
"tests/**" = ["S101"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=ocr_webapp --cov-report=term-missing --cov-report=html"

[tool.coverage.run]
omit = ["tests/*", "scripts/*"]
